✨ feat(policies): add support for history policies access
- add `HISTORY_ACCESS_POLICIES_GROUPS` environment variable
- update casl ability factory to handle policy access groups
- extend configuration to include history policies groups

♻️ refactor(policies): enhance policy service and module

- refactor policies service to use $set and run validators
- convert policy to async factory to add history plugin
- add logging for policy removal process

📝 docs: update README with new environment variable

- document `HISTORY_ACCESS_POLICIES_GROUPS` in README
- explain its role in accessing history policies data records

+++ src/policies/policies.service.ts
@@ -138,14 +138,34 @@ export class PoliciesService implements OnModuleInit {
   ): Promise<Policy | null> {
     const username = (this.request.user as JWTUser).username;
     return this.policyModel
-      .findOneAndUpdate(filter, addUpdatedByField(updatePolicyDto, username), {
-        new: true,
-      })
+      .findOneAndUpdate(
+        filter,
+        {
+          $set: {
+            ...updatePolicyDto,
+            updatedBy: username,
+            updatedAt: new Date(),
+          },
+        },
+        {
+          new: true,
+          runValidators: true,
+        },
+      )
       .exec();
   }
 
   async remove(filter: FilterQuery<PolicyDocument>): Promise<unknown> {
-    return this.policyModel.findOneAndDelete(filter).exec();
+    console.log("Removing policy with filter:", filter);
+    const existingDoc = await this.policyModel.findOne(filter).exec();
+
+    if (!existingDoc) {
+      console.log("No document found to remove with filter:", filter);
+      return null;
+    }
+    const result = await this.policyModel.findOneAndDelete(filter).exec();
+
+    return result;
   }
 
   async updateWhere(ownerGroupList: string, data: UpdatePolicyDto) {
